+++ backend/bots/examples/satoshi_dice/c2c_client/src/lib.rs
@@ -1,0 +1,7 @@
+use canister_client::generate_c2c_call;
+use satoshi_dice_canister::*;
+
+// Queries
+
+// Updates
+generate_c2c_call!(c2c_add_user);