+++ backend/canisters/group/impl/src/lib.rs
@@ -184,6 +184,7 @@ impl RuntimeState {
             public: self.data.is_public,
             date_created: self.data.date_created,
             members: self.data.participants.len(),
+            moderators: self.data.participants.moderator_count(),
             admins: self.data.participants.admin_count(),
             owners: self.data.participants.owner_count(),
             blocked: self.data.participants.blocked().len() as u32,
@@ -334,12 +335,11 @@ impl Data {
         }
     }
 
-    pub fn is_accessible_by_non_member(&self, caller: Principal, invite_code: Option<u64>) -> bool {
-        if self.is_public || self.invited_users.get(&caller).is_some() {
-            return true;
-        }
-
-        self.is_invite_code_valid(invite_code)
+    pub fn is_accessible(&self, caller: Principal, invite_code: Option<u64>) -> bool {
+        self.is_public
+            || self.participants.get_by_principal(&caller).is_some()
+            || self.invited_users.get(&caller).is_some()
+            || self.is_invite_code_valid(invite_code)
     }
 
     pub fn is_frozen(&self) -> bool {
@@ -369,6 +369,7 @@ pub struct Metrics {
     pub public: bool,
     pub date_created: TimestampMillis,
     pub members: u32,
+    pub moderators: u32,
     pub admins: u32,
     pub owners: u32,
     pub blocked: u32,