+++ backend/canisters/group/impl/src/updates/mod.rs
@@ -1,4 +1,3 @@
-mod add_participants;
 mod add_reaction;
 mod c2c_delete_group;
 mod c2c_freeze_group;
@@ -25,7 +24,7 @@ mod register_proposal_vote;
 mod register_proposal_vote_v2;
 mod remove_participant;
 mod remove_reaction;
-mod send_message_v2;
+mod send_message;
 mod unblock_user;
 mod undelete_messages;
 mod unpin_message;