+++ backend/integration_tests/src/client/group.rs
@@ -11,7 +11,6 @@ generate_query_call!(summary);
 generate_query_call!(summary_updates);
 
 // Updates
-generate_update_call!(add_participants);
 generate_update_call!(add_reaction);
 generate_update_call!(block_user);
 generate_update_call!(change_role);
@@ -34,27 +33,8 @@ pub mod happy_path {
     use ic_test_state_machine_client::StateMachine;
     use types::{
         ChatId, EventIndex, EventsResponse, GroupCanisterGroupChatSummary, GroupCanisterGroupChatSummaryUpdates,
-        MessageContentInitial, MessageId, MessageIndex, PollVotes, TextContent, TimestampMillis, UserId, VoteOperation,
+        MessageContentInitial, MessageId, MessageIndex, PollVotes, TextContent, TimestampMillis, VoteOperation,
     };
-
-    pub fn add_participants(env: &mut StateMachine, sender: &User, group_chat_id: ChatId, user_ids: Vec<UserId>) {
-        let response = super::add_participants(
-            env,
-            sender.principal,
-            group_chat_id.into(),
-            &group_canister::add_participants::Args {
-                user_ids,
-                added_by_name: sender.username(),
-                allow_blocked_users: false,
-                correlation_id: 0,
-            },
-        );
-
-        match response {
-            group_canister::add_participants::Response::Success => {}
-            response => panic!("'add_participants' error: {response:?}"),
-        }
-    }
 
     pub fn send_text_message(
         env: &mut StateMachine,
@@ -81,7 +61,7 @@ pub mod happy_path {
         );
 
         match response {
-            group_canister::send_message::Response::Success(result) => result,
+            group_canister::send_message_v2::Response::Success(result) => result,
             response => panic!("'send_message' error: {response:?}"),
         }
     }