+++ backend/integration_tests/src/client/local_user_index.rs
@@ -4,6 +4,7 @@ use local_user_index_canister::*;
 // Queries
 
 // Updates
+generate_update_call!(invite_users_to_group);
 generate_update_call!(join_group);
 generate_update_call!(register_user);
 generate_update_call!(report_message);
@@ -14,7 +15,7 @@ pub mod happy_path {
     use crate::User;
     use candid::Principal;
     use ic_test_state_machine_client::StateMachine;
-    use types::{CanisterId, ChatId};
+    use types::{CanisterId, ChatId, UserId};
 
     pub fn register_user(env: &mut StateMachine, canister_id: CanisterId) -> User {
         register_user_with_referrer(env, canister_id, None)
@@ -45,6 +46,30 @@ pub mod happy_path {
         }
     }
 
+    pub fn invite_users_to_group(
+        env: &mut StateMachine,
+        sender: Principal,
+        local_user_index_canister_id: CanisterId,
+        group_id: ChatId,
+        user_ids: Vec<UserId>,
+    ) {
+        let response = super::invite_users_to_group(
+            env,
+            sender,
+            local_user_index_canister_id,
+            &local_user_index_canister::invite_users_to_group::Args {
+                group_id,
+                user_ids,
+                correlation_id: 0,
+            },
+        );
+
+        match response {
+            local_user_index_canister::invite_users_to_group::Response::Success => {}
+            response => panic!("'invite_users_to_group' error: {response:?}"),
+        }
+    }
+
     pub fn join_group(env: &mut StateMachine, sender: Principal, local_user_index_canister_id: CanisterId, chat_id: ChatId) {
         let response = super::join_group(
             env,
@@ -61,5 +86,27 @@ pub mod happy_path {
             local_user_index_canister::join_group::Response::Success(_) => {}
             response => panic!("'join_group' error: {response:?}"),
         }
+    }
+
+    pub fn add_users_to_group(
+        env: &mut StateMachine,
+        sender: Principal,
+        local_user_index_canister_id: CanisterId,
+        group_id: ChatId,
+        users: Vec<(UserId, Principal)>,
+    ) {
+        invite_users_to_group(
+            env,
+            sender,
+            local_user_index_canister_id,
+            group_id,
+            users.iter().map(|(user_id, _)| *user_id).collect(),
+        );
+
+        for (_, principal) in users {
+            join_group(env, principal, local_user_index_canister_id, group_id);
+        }
+
+        env.tick();
     }
 }