+++ frontend/app/src/components/Router.svelte
@@ -16,6 +16,8 @@
         page("/whitepaper", parsePathParams, track, () => (route = LandingPage));
         page("/miami", parsePathParams, track, () => (route = LandingPage));
         page("/guidelines", parsePathParams, track, () => (route = LandingPage));
+        page("/faq", parsePathParams, track, () => (route = LandingPage));
+        page("/diamond", parsePathParams, track, () => (route = LandingPage));
         page("/architecture", parsePathParams, track, () => (route = LandingPage));
         page(
             "/:chatId?/:messageIndex?/:threadMessageIndex?",