+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Register new Bitcoin Miami users with the SatoshiDice bot ([#3591](https://github.com/open-ic/open-chat/pull/3591))
+
 ## [[2.0.687](https://github.com/open-ic/open-chat/releases/tag/v2.0.687-local_user_index)] - 2023-05-12
 
 ### Added